move GetJsonSchemaHandler and GetCoreSchemaHandler to be properly public
