+++ pydantic/__init__.py
@@ -1,4 +1,4 @@
-from pydantic_core import ValidationError
+import pydantic_core
 from pydantic_core.core_schema import (
     FieldSerializationInfo,
     FieldValidationInfo,
@@ -26,6 +26,9 @@ from .validate_call import validate_call
 from .version import VERSION
 
 __version__ = VERSION
+
+# this encourages pycharm to import `ValidationError` from here, not pydantic_core
+ValidationError = pydantic_core.ValidationError
 
 # WARNING __all__ from .errors is not included here, it will be removed as an export here in v2
 # please use "from pydantic.errors import ..." instead
