+++ pydantic/_internal/_schema_generation_shared.py
@@ -3,59 +3,18 @@ Types and utility functions used by various other internal tools.
 """
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Callable, Dict
+from typing import TYPE_CHECKING, Any, Callable
 
 from pydantic_core import core_schema
 
-from pydantic._internal._core_utils import CoreSchemaOrField
+from ..annotated import GetCoreSchemaHandler, GetJsonSchemaHandler
 
 if TYPE_CHECKING:
-    from pydantic.json_schema import GenerateJsonSchema, JsonSchemaMode
+    from ..json_schema import GenerateJsonSchema, JsonSchemaValue
+    from ._core_utils import CoreSchemaOrField
 
-JsonSchemaValue = Dict[str, Any]
-
-
-class GetJsonSchemaHandler:
-    """
-    Handler to call into the next JSON schema generation function
-    """
-
-    mode: JsonSchemaMode
-
-    def __call__(self, __core_schema: CoreSchemaOrField) -> JsonSchemaValue:
-        """Call the inner handler and get the JsonSchemaValue it returns.
-        This will call the next JSON schema modifying function up until it calls
-        into `pydantic.json_schema.GenerateJsonSchema`, which will raise a
-        `pydantic.errors.PydanticInvalidForJsonSchema` error if it cannot generate
-        a JSON schema.
-
-        Args:
-            __core_schema (CoreSchemaOrField): A `pydantic_core.core_schema.CoreSchema`.
-
-        Returns:
-            JsonSchemaValue: the JSON schema generated by the inner JSON schema modify
-            functions.
-        """
-        raise NotImplementedError
-
-    def resolve_ref_schema(self, __maybe_ref_json_schema: JsonSchemaValue) -> JsonSchemaValue:
-        """Get the real schema for a `{"$ref": ...}` schema.
-        If the schema given is not a `$ref` schema, it will be returned as is.
-        This means you don't have to check before calling this function.
-
-        Args:
-            __maybe_ref_json_schema (JsonSchemaValue): A JsonSchemaValue, ref based or not.
-
-        Raises:
-            LookupError: if the ref is not found.
-
-        Returns:
-            JsonSchemaValue: A JsonSchemaValue that has no `$ref`.
-        """
-        raise NotImplementedError
-
-
-GetJsonSchemaFunction = Callable[[CoreSchemaOrField, GetJsonSchemaHandler], JsonSchemaValue]
+    GetJsonSchemaFunction = Callable[[CoreSchemaOrField, GetJsonSchemaHandler], JsonSchemaValue]
+    HandlerOverride = Callable[[CoreSchemaOrField], JsonSchemaValue]
 
 
 class UnpackedRefJsonSchemaHandler(GetJsonSchemaHandler):
@@ -64,7 +23,7 @@ class UnpackedRefJsonSchemaHandler(GetJsonSchemaHandler):
     schemas so that the caller doesn't have to worry about that.
 
     This is used for custom types and models that implement `__get_pydantic_core_schema__`
-    so they they always get a `non-$ref` schema.
+    so they always get a `non-$ref` schema.
 
     Used internally by Pydantic, please do not rely on this implementation.
     See `GetJsonSchemaHandler` for the handler API.
@@ -111,9 +70,6 @@ def wrap_json_schema_fn_for_model_or_custom_type_with_ref_unpacking(
     return wrapped
 
 
-HandlerOverride = Callable[[CoreSchemaOrField], JsonSchemaValue]
-
-
 class GenerateJsonSchemaHandler(GetJsonSchemaHandler):
     """
     JsonSchemaHandler implementation that doesn't do ref unwrapping by default.
@@ -144,44 +100,6 @@ class GenerateJsonSchemaHandler(GetJsonSchemaHandler):
                 ' Maybe you tried to call resolve_ref_schema from within a recursive model?'
             )
         return json_schema
-
-
-class GetCoreSchemaHandler:
-    """
-    Handler to call into the next CoreSchema schema generation function
-    """
-
-    def __call__(self, __source_type: Any) -> core_schema.CoreSchema:
-        """
-        Call the inner handler and get the CoreSchema it returns.
-        This will call the next CoreSchema modifying function up until it calls
-        into Pydantic's internal schema generation machinery, which will raise a
-        `pydantic.errors.PydanticSchemaGenerationError` error if it cannot generate
-        a CoreSchema for the given source type.
-
-        Args:
-            __source_type (Any): The input type.
-
-        Returns:
-            CoreSchema: the `pydantic-core` CoreSchema generated.
-        """
-        raise NotImplementedError
-
-    def generate_schema(self, __source_type: Any) -> core_schema.CoreSchema:
-        """
-        Generate a schema unrelated to the current context.
-        Use this function if e.g. you are handling schema generation for a sequence
-        and want to generate a schema for it's items.
-        Otherwise you may end up doing something like applying a `min_length` constraint
-        that was intended for the sequence itself to it's items!
-
-        Args:
-            __source_type (Any): The input type.
-
-        Returns:
-            CoreSchema: the `pydantic-core` CoreSchema generated.
-        """
-        raise NotImplementedError
 
 
 class CallbackGetCoreSchemaHandler(GetCoreSchemaHandler):
